Web_Application/Server/requirements.txt
flask==3.0.0
ortools==9.7.2996
osmnx==1.3.0
networkx==3.1
folium==0.14.0
requests==2.31.0
python-dotenv==1.0.0
geopy==2.3.0
pandas==2.0.3